--- WpfAnalyzers.Test/WPF0080MarkupExtensionDoesNotHaveAttributeTests/InValidCode.cs
@@ -1,38 +1,0 @@
-namespace WpfAnalyzers.Test.WPF0080MarkupExtensionDoesNotHaveAttributeTests
-{
-    using Gu.Roslyn.Asserts;
-    using Microsoft.CodeAnalysis.Diagnostics;
-    using NUnit.Framework;
-
-    public class InValidCode
-    {
-        private static readonly DiagnosticAnalyzer Analyzer = new WPF0080MarkupExtensionDoesNotHaveAttribute();
-
-        [Test]
-        public void WhenIsGeneric()
-        {
-            var testCode = @"
-namespace RoslynSandbox
-{
-    using System;
-    using System.Windows.Markup;
-
-    public abstract class ConverterMarkupExtension<T> : MarkupExtension, IValueConverter
-        where T : class, new()
-    {
-        private static T _converter;
-
-        public override object ProvideValue(IServiceProvider serviceProvider)
-        {
-            return _converter ?? (_converter = new T());
-        }
-
-        public abstract object Convert(object value, Type targetType, object parameter, CultureInfo culture);
-
-        public abstract object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture);
-    }
-}";
-            RoslynAssert.NoAnalyzerDiagnostics(Analyzer, testCode);
-        }
-    }
-}
