+++ WpfAnalyzers.Test/WPF0080MarkupExtensionDoesNotHaveAttributeTests/ValidCode.cs
@@ -66,5 +66,34 @@ namespace RoslynSandbox
 }";
             RoslynAssert.Valid(Analyzer, testCode);
         }
+
+        [Test]
+        public void WhenIsGeneric()
+        {
+            var testCode = @"
+namespace RoslynSandbox
+{
+    using System;
+    using System.Globalization;
+    using System.Windows.Data;
+    using System.Windows.Markup;
+
+    public abstract class ConverterMarkupExtension<T> : MarkupExtension, IValueConverter
+        where T : class, new()
+    {
+        private static T _converter;
+
+        public override object ProvideValue(IServiceProvider serviceProvider)
+        {
+            return _converter ?? (_converter = new T());
+        }
+
+        public abstract object Convert(object value, Type targetType, object parameter, CultureInfo culture);
+
+        public abstract object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture);
+    }
+}";
+            RoslynAssert.Valid(Analyzer, testCode);
+        }
     }
 }
